chore: rename aws boilerplate
